Student and Course CRUD Action Menthods and Views Added

+++ WebProgramming/OlineStore/Controllers/HomeController.cs
@@ -1,6 +1,9 @@
-﻿using OnlineStore.Models;
+﻿using OnlineStore.Extensions;
+using OnlineStore.Models;
 using System.Linq;
 using System.Web.Mvc;
+using System.Web.Routing;
+using System.Web.UI.WebControls;
 
 namespace OnlineStore.Controllers
 {
@@ -18,10 +21,17 @@ namespace OnlineStore.Controllers
             //return View(products);
 
             var students = dbContext.Students.ToList();
-            
+            //var filterd = from s in students
+            //              where s.StudentId < 5
+            //              select s;
+
+            //var filter2 = students.Where(s => s.StudentId < 4).ToList();
+        
             return View(students);
         }
 
+
+        //Student Action Methods
         [HttpGet]
         [Route("student/create")]
         public ActionResult Create()
@@ -33,32 +43,45 @@ namespace OnlineStore.Controllers
         [Route("student/create")]
         public ActionResult Create(Student student)
         {
-            dbContext.Students.Add(student);
-            dbContext.SaveChanges();
+            if (ModelState.IsValid)
+            {
+                dbContext.Students.Add(student);
+                dbContext.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
 
         [HttpGet]
-        [Route("student/edit/{id}")]
+        [Route("student/edit/{id:int}")]
         public ActionResult Edit(int? id)
         {
             if (id.HasValue)
             {
-                var student = dbContext.Students.FirstOrDefault(s=> s.StudentId == id);
+                //var student = dbContext.Students.FirstOrDefault(s=> s.StudentId == id);
+                var student = dbContext.Students.Find(id);
+
+                //Only purpose is to testing the edit view  
+                student.AddDefaultCourses();
+
                 return View(student);
             }
             return RedirectToAction("Index");
         }
+
         [HttpPost]
         [Route("student/edit")]
         public ActionResult Edit(Student student) 
         {
-            if(student == null) return View();
-            dbContext.Entry(student).State = System.Data.Entity.EntityState.Modified; 
-            dbContext.SaveChanges(); 
-            return RedirectToAction("Index");    
+            if (ModelState.IsValid)
+            {
+                if (student == null) return View();
+                dbContext.Entry(student).State = System.Data.Entity.EntityState.Modified;
+                dbContext.SaveChanges();
+            }
+            return RedirectToAction("Index");
         }
 
+      
         [HttpGet]
         [Route("student/delete/{id}")]
         public ActionResult Delete(int id)
@@ -70,15 +93,55 @@ namespace OnlineStore.Controllers
         [Route("student/delete")]
         public ActionResult Delete(Student student)
         {
-            var toBeRemovedStudent = dbContext.Students.Find(student.StudentId);
-            if(toBeRemovedStudent != null)
+            if (ModelState.IsValid)
             {
-                dbContext.Students.Remove(toBeRemovedStudent);
-                dbContext.SaveChanges();
-                return RedirectToAction("Index");
+                var toBeRemovedStudent = dbContext.Students.Find(student.StudentId);
+                if (toBeRemovedStudent != null)
+                {
+                    dbContext.Students.Remove(toBeRemovedStudent);
+                    dbContext.SaveChanges();
+                    return RedirectToAction("Index");
+                }
             }
             return new HttpStatusCodeResult(System.Net.HttpStatusCode.Conflict);
         }
+
+        //Course Action Methods
+        [HttpGet]
+        [Route("course/edit/{studentId:int}/{courseId:int}")]
+        public ActionResult EditCourse(int studentId, int courseId)
+        {
+            var student = dbContext.Students.Find(studentId);
+            student.AddDefaultCourses();
+            
+            var course = student.Courses.FirstOrDefault(c => c.CourseId == courseId);
+
+            course.Student = student;
+
+            return View(course);
+        }
+
+        [HttpPost]
+        [Route("course/edit")]
+        public ActionResult EditCourse(Course course)
+        {
+            //if(ModelState.IsValid)
+            //{
+            //    dbContext.Entry(course).State = System.Data.Entity.EntityState.Modified;
+            //    dbContext.SaveChanges();
+            //}
+            //return RedirectToRoute(new {controller="home", action="Edit" });
+            
+            return Redirect("/student/edit/1");
+        }
+
+        [HttpGet]
+        [Route("course/delete/{id:int}")]
+        public ActionResult DeleteCourse()
+        {
+            return Content("<h2>The Delete Action Method is not implemented</h2>");
+        }
+
 
         public ActionResult About()
         {
